+++ source/module_hsolver/test/hsolver_supplementary_mock.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "module_elecstate/elecstate.h"
+#include "module_elecstate/elecstate_pw.h"
 #include "module_psi/wavefunc.h"
 
 namespace elecstate
@@ -61,6 +61,46 @@ void ElecState::init_ks(Charge* chg_in, // pointer for class Charge
 {
     return;
 }
+
+template <typename T, typename Device>
+ElecStatePW<T, Device>::ElecStatePW(ModulePW::PW_Basis_K* wfc_basis_in,
+                                    Charge* chg_in,
+                                    K_Vectors* pkv_in,
+                                    UnitCell* ucell_in,
+                                    pseudopot_cell_vnl* ppcell_in,
+                                    ModulePW::PW_Basis* rhodpw_in,
+                                    ModulePW::PW_Basis* rhopw_in,
+                                    ModulePW::PW_Basis_Big* bigpw_in)
+    : basis(wfc_basis_in)
+{
+}
+
+template <typename T, typename Device>
+ElecStatePW<T, Device>::~ElecStatePW()
+{
+}
+
+template <typename T, typename Device>
+void ElecStatePW<T, Device>::psiToRho(const psi::Psi<T, Device>& psi)
+{
+}
+
+template <typename T, typename Device>
+void ElecStatePW<T, Device>::cal_tau(const psi::Psi<T, Device>& psi)
+{
+}
+
+template <typename T, typename Device>
+void ElecStatePW<T, Device>::cal_becsum(const psi::Psi<T, Device>& psi)
+{
+}
+
+template class ElecStatePW<std::complex<float>, base_device::DEVICE_CPU>;
+template class ElecStatePW<std::complex<double>, base_device::DEVICE_CPU>;
+#if ((defined __CUDA) || (defined __ROCM))
+template class ElecStatePW<std::complex<float>, base_device::DEVICE_GPU>;
+template class ElecStatePW<std::complex<double>, base_device::DEVICE_GPU>;
+#endif
 
 Potential::~Potential()
 {
